backend finalizado

+++ tareas/backend/controllers/tareasController.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/treaModel')
 
 
 const getTareas = asyncHandler (async (req,res) => {
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({user: req.user._id})
     res.status(200).json(tareas)    
 })
 
@@ -14,6 +14,7 @@ const createTarea = asyncHandler (async (req,res) => {
     }
     const tarea = await Tarea.create({
         texto: req.body.texto,
+        user: req.user._id
     })
     res.status(201).json(tarea)        
 })
@@ -24,9 +25,15 @@ const updateTarea = asyncHandler (async (req,res) => {
         res.status(404)
         throw new Error ('tarea no enconrtada')
     }
+
+    //Verifica,ps que la tarea le pertenece al usuario del token proporcionado 
+    if(tarea.user.toString() !== req.user._id){
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    }else{
     const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new:true})
     res.status(200).json(updatedTarea)
-
+    }
 })
 
 const deleteTarea = asyncHandler (async (req,res) => {
@@ -35,8 +42,14 @@ const deleteTarea = asyncHandler (async (req,res) => {
         res.status(404)
         throw new Error ('tarea no enconrtada')
     }
-    await Tarea.deleteOne(tarea)
-    res.status(200).json({id: req.params.id})
+    if(tarea.user.toString() !== req.user._id){
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    }else{
+        await Tarea.deleteOne(tarea)
+        res.status(200).json({id: req.params.id})
+    }
+  
 })
 
 module.exports = {
